[Impeller] Fix blown stack due to always out of date window dimensions on high-dpi devices. (#50307)
This regression was introduced in https://github.com/flutter/engine/commit/c25f3f5e65c0c506e60ae664aff79bdfa762f85b which made it so that the updated surface size would be provided to the swapchain impl out-of-band instead of being inferred from the surface properties (to avoid a lag between the reports of the two).

However, the playgrounds used the window size to report the updated size. This is fine where the window scale is 1. But fails on high-dpi devices. The patch replaces the call to get the window size with an equivalent call to get the framebuffer size.

The failure was a blown stack due to a new swapchain impl being created over and over again because the size was always out of date.

Fixes https://github.com/flutter/flutter/issues/142833